requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic>=2.6.0
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.24.0
pymongo>=4.5.0
dnspython>=2.4.0
Pillow>=10.0.0
